+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0).
 
 ## Unreleased
 
-Nothing.
+### Added
+
+- Ensure that at least one filter is provided to script (in any mode).
 
 ## [1.0.0](https://github.com/trallnag/filter-pre-commit-hooks/compare/6014a859fec1b8842ea1dc573e096609e61ceecd...v1.0.0) / 2025-01-01